+++ test/controllers/application_controller_test.rb
@@ -4,14 +4,6 @@ require 'test_helper'
 
 # Tests to assure display of static pages.
 class ApplicationControllerTest < ActionController::TestCase
-  test 'should parse time' do
-    assert_equal '12:00:00', @controller.send(:parse_time_to_s, '12:00:00')
-  end
-
-  test 'should parse invalid time' do
-    assert_equal '', @controller.send(:parse_time_to_s, 'abc')
-  end
-
   test 'should get version' do
     get :version
     assert_response :success